spendmate-ocr/requirements.txt
flask
paddlepaddle
paddleocr
pillow
python-dotenv
requests
